Latest version for all packages, implemented database framework

+++ src/database/models.rs
@@ -1,14 +1,14 @@
 #[derive(Queryable, Debug)]
 pub struct User {
   pub id: i32,
-  pub discordid: i32,
+  pub discordid: i64,
   pub role: String,
 }
 
 #[derive(Insertable)]
 #[table_name = "users"]
 pub struct NewUser<'a> {
-  pub discordid: i32,
+  pub discordid: i64,
   pub role: &'a str,
 }
